// src/types/custom-d-t.ts
export type CartItemId = string | number;

export interface CartItem {
  id: CartItemId;
  title: string;
  price: number;
  image: import("next/image").StaticImageData | string;
  link?: string;
  quantity: number;
}

export interface AppContextType {
  openOffcanvas: boolean;
  setOpenOffcanvas: React.Dispatch<React.SetStateAction<boolean>>;
  openCartOffcanvas: boolean;
  setOpenCartOffcanvas: React.Dispatch<React.SetStateAction<boolean>>;
  toggleCartOffcanvas: () => void;
  toggleOffcanvas: () => void;
  toggleSearch: () => void;
  toggleModal: () => void;
  openModal: boolean;
  openSearch: boolean;
  setOpenSearch: React.Dispatch<React.SetStateAction<boolean>>;
  // --- cart ---
  cartItems: CartItem[];
  addToCart: (item: CartItem) => void;
  removeFromCart: (id: CartItemId) => void;
  incQty: (id: CartItemId) => void;
  decQty: (id: CartItemId) => void;
  clearCart: () => void;
  cartCount: number;
  cartTotal: number;
}